moving the index blog pages to the templates
