+++ src/pages/blog/index.en.md
@@ -1,0 +1,11 @@
+---
+id: '02'
+templateKey: blog-index-en
+title: Blog index page.
+date: 2019-03-24
+description: Index page for all blog posts.
+lang: en
+path: /en/blog/
+tags:
+  - art
+---
