+++ src/pages/blog/index.it.md
@@ -1,0 +1,11 @@
+---
+id: '02'
+templateKey: blog-index-it
+title: Pagina indice blog.
+date: 2019-03-24
+description: Pagina indice per tutti i blog posts.
+lang: it
+path: /it/blog/
+tags:
+  - art
+---
