+++ src/templates/home.js
@@ -60,16 +60,6 @@ export const pageQuery = graphql`
         }
       }
     }
-    allJavascriptFrontmatter {
-  edges {
-    node {
-      frontmatter {
-        id
-        title
-      }
-    }
-  }
-}
     markdownRemark(id: {eq: $id}) {
       html
       frontmatter {
